// dfp-piao-fe/src/app/shared/models/enums/cod-path.enum.ts
export enum CodPathEnum {
  SEZIONE_1 = 'sezione1',
  SEZIONE_2_1 = 'sezione21',
  SEZIONE_2_2 = 'sezione22',
  SEZIONE_2_3 = 'sezione23',
  SEZIONE_3_1 = 'sezione31',
  SEZIONE_3_2 = 'sezione32',
  SEZIONE_3_3_1 = 'sezione331',
  SEZIONE_3_3_2 = 'sezione332',
  SEZIONE_4 = 'sezione4',
  PIAO = 'piao',
  TICKET = 'ticket',
}